requirements.txt
asgiref==3.5.2
Babel==2.11.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.2
django-money==3.0.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
py-moneyed==2.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
